feat : Sort logic util 함수화
